// src/types/product.ts
export type Product = {
  image: string;
  productName: string;
  category: string;
  price: number;
  sold: number;
  profit: number;
  brandName: string; // Thêm thuộc tính brandName
  description: string;
  id: number;
  weight: number;
  code: string;
  unit: string;
  quantity: number;
  quantityReturn: number;
  idBath: number;
  discount: number;
  note: String;
  batch: String;
  location: string;
};